Solved deprecated warnings in test utils

+++ simpa_tests/test_utils/__init__.py
@@ -25,10 +25,10 @@ def assert_equals_recursive(a, b):
                     assert (a[item] == b[item]).all()
                 else:
                     assert a[item] == b[item], str(a[item]) + " is not the same as " + str(b[item])
-                    if (np.issubdtype(type(a[item]), np.int) or np.issubdtype(type(a[item]), np.float)) and \
-                            (np.issubdtype(type(b[item]), np.int) or np.issubdtype(type(b[item]), np.float)):
-                        assert (np.issubdtype(type(a[item]), np.int) & np.issubdtype(type(b[item]), np.int)) or \
-                               (np.issubdtype(type(a[item]), np.float) & np.issubdtype(type(b[item]), np.float))
+                    if (np.issubdtype(type(a[item]), int) or np.issubdtype(type(a[item]), float)) and \
+                            (np.issubdtype(type(b[item]), int) or np.issubdtype(type(b[item]), float)):
+                        assert (np.issubdtype(type(a[item]), int) & np.issubdtype(type(b[item]), int)) or \
+                               (np.issubdtype(type(a[item]), float) & np.issubdtype(type(b[item]), float))
                     else:
                         assert isinstance(a[item], type(b[item]))
                         assert isinstance(b[item], type(a[item]))
